pr edits made, all tests passing
