+++ packages/contracts/package.json
@@ -12,8 +12,8 @@
     "help": "hardhat help",
     "clean": "hardhat clean",
     "lint": "solhint contracts/{*,**/*}.sol",
-    "x": "hardhat test ./test/SmartInvoiceFactoryV2.js",
-    "y": "hardhat test ./test/SmartInvoiceEscrowV2.js",
+    "x": "hardhat test ./test/SmartInvoiceFactory.js",
+    "y": "hardhat test ./test/SmartInvoiceEscrow.js",
     "deploy-local": "hardhat run scripts/deploy-factory-local.js --network localhost",
     "deploy-kovan": "hardhat run scripts/deploy-factory.js --network kovan",
     "deploy-mainnet": "hardhat run scripts/deploy-factory.js --network mainnet",
