+++ packages/contracts/test/utils.js
@@ -7,7 +7,7 @@ const EMPTY_BYTES32 =
 module.exports.awaitInvoiceAddress = async receipt => {
   if (!receipt || !receipt.logs) return "";
   const abi = new ethers.utils.Interface([
-    "event LogNewInvoice(uint256 indexed id, address invoice, uint256[] amounts)",
+    "event LogNewInvoice(uint256 indexed id, address invoice, uint256[] amounts, bytes32 implementationType,uint256 implementationVersion, address implementationAddress)",
   ]);
   const eventFragment = abi.events[Object.keys(abi.events)[0]];
   const eventTopic = abi.getEventTopic(eventFragment);
@@ -28,45 +28,45 @@ module.exports.currentTimestamp = async () => {
   return +block.timestamp;
 };
 
-module.exports.getLockedInvoice = async (
-  SmartInvoice,
-  client,
-  provider,
-  resolverType,
-  resolver,
-  mockToken,
-  amounts,
-  resolutionRate,
-  details,
-  mockWrappedNativeToken,
-  value = 0,
-) => {
-  const currentTime = await module.exports.currentTimestamp();
-  const newInvoice = await SmartInvoice.deploy();
-  await newInvoice.deployed();
-  await newInvoice.init(
-    client.address,
-    provider.address,
-    resolverType,
-    resolver.address,
-    mockToken.address,
-    amounts,
-    currentTime + 1000,
-    resolutionRate,
-    details,
-    mockWrappedNativeToken.address,
-    false,
-  );
-  expect(await newInvoice["locked()"]()).to.equal(false);
-  await mockToken.mock.balanceOf.withArgs(newInvoice.address).returns(10);
-  const receipt = newInvoice["lock(bytes32)"](EMPTY_BYTES32, { value });
-  await expect(receipt)
-    .to.emit(newInvoice, "Lock")
-    .withArgs(client.address, EMPTY_BYTES32);
-  return newInvoice;
-};
+// module.exports.getLockedInvoice = async (
+//   SmartInvoice,
+//   client,
+//   provider,
+//   resolverType,
+//   resolver,
+//   mockToken,
+//   amounts,
+//   resolutionRate,
+//   details,
+//   mockWrappedNativeToken,
+//   value = 0,
+// ) => {
+//   const currentTime = await module.exports.currentTimestamp();
+//   const newInvoice = await SmartInvoice.deploy();
+//   await newInvoice.deployed();
+//   await newInvoice.init(
+//     client.address,
+//     provider.address,
+//     resolverType,
+//     resolver.address,
+//     mockToken.address,
+//     amounts,
+//     currentTime + 1000,
+//     resolutionRate,
+//     details,
+//     mockWrappedNativeToken.address,
+//     false,
+//   );
+//   expect(await newInvoice["locked()"]()).to.equal(false);
+//   await mockToken.mock.balanceOf.withArgs(newInvoice.address).returns(10);
+//   const receipt = newInvoice["lock(bytes32)"](EMPTY_BYTES32, { value });
+//   await expect(receipt)
+//     .to.emit(newInvoice, "Lock")
+//     .withArgs(client.address, EMPTY_BYTES32);
+//   return newInvoice;
+// };
 
-module.exports.getLockedInvoiceV2 = async (
+module.exports.getLockedInvoice = async (
   SmartInvoiceEscrowV2,
   client,
   provider,
@@ -74,7 +74,7 @@ module.exports.getLockedInvoiceV2 = async (
   amounts,
   mockWrappedNativeToken,
   implementationData,
-  implementationInfoData,
+  invoiceId,
   mockToken,
   value = 0,
 ) => {
@@ -87,7 +87,7 @@ module.exports.getLockedInvoiceV2 = async (
     amounts,
     mockWrappedNativeToken.address,
     implementationData,
-    implementationInfoData,
+    invoiceId,
   );
   expect(await newInvoice["locked()"]()).to.equal(false);
   await mockToken.mock.balanceOf.withArgs(newInvoice.address).returns(10);
